+++ src/main/java/ua/training/servlet/hospital/dao/DiagnosisDao.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface DiagnosisDao extends GenericDao<Diagnosis> {
     List<Diagnosis> findDiagnosesByPatientId(int start, int count, long patientId);
-    int countDiagnosesOfPatient(long patientId);
+    long countDiagnosesOfPatient(long patientId);
 }
